+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v0.29.1 (master)
 * Fix #1381 Use the correct peer_name when initiating the TLS connection
 * First changes for upcoming PHP 8.4 support
 * Fix #1386 Do a manual query to delete the subscription to prevent using the composite primary key in Eloquent
+* Close the StoriesViewer using the Esc button or by clicking outside the Story
 
 v0.29
 ---------------------------