+++ CHANGELOG.md
@@ -12,6 +12,7 @@ v0.29.1 (master)
 * Close the StoriesViewer using the Esc button or by clicking outside the Story
 * Add mucjid in the Presence "primary key", an actual key for PostgreSQL and an unique hash column for MySQL (3072 key length limit)
 * Fix #1357 Rewrite the PresenceBufferSaver delete request that was not generating the correct SQL code (using OR and not AND in it)
+* Update the translations
 
 v0.29
 ---------------------------